revise lyric crawler
