+++ daemon/provider/lyric/lyricprovider.py
@@ -1,7 +1,3 @@
-from .metroprovider import MetroProvider
-from .geniusprovider import GeniusProvider
-
-
 class LyricProvider:
 
     def get_lyric(self, artist_name, track_name):
